% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{accessions.mor}
\alias{accessions.mor}
\title{Morton Arboretum living collections accessions, 2021-06-22}
\format{
a \code{data.frame} of 2173 observations, 14 variables:
}
\usage{
data(accessions.mor)
}
\description{
A dataset containing living collections accessions for \emph{Malus}, \emph{Quercus},
\emph{Tilia} and \emph{Ulmus} housed at The Morton Arboretum. Data downloaded from
http://bol.mortonarb.org 2021-06-22. The variables are as follows:
}
\details{
\itemize{
\item Taxon.
\item Vernacular.
\item Trademark.
\item ProvenanceType.
\item DonorInstitute.
\item CountryName.
\item PrincipalCollector.
\item LocationSummary.
\item BgLocation.
\item GardenSubarea.
\item GardenGrid.
\item GardenSubGrid.
\item CommemorationText.
}
}
\keyword{datasets}
