% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed.calc.R
\name{ed.calc}
\alias{ed.calc}
\title{Calculate evolutionary distinctiveness
Calculates evolutionary distinctiveness for trees in a \code{consInt} object}
\usage{
ed.calc(x)
}
\arguments{
\item{x}{a \code{consInt} object}
}
\description{
Calculate evolutionary distinctiveness
Calculates evolutionary distinctiveness for trees in a \code{consInt} object
}
