% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eDisCalc.R
\name{eDisCalc}
\alias{eDisCalc}
\title{Calculate evolutionary distinctiveness
Calculates evolutionary distinctiveness for trees in a \code{consInt} object}
\usage{
eDisCalc(x)
}
\arguments{
\item{x}{a \code{consInt} object}
}
\description{
Calculate evolutionary distinctiveness
Calculates evolutionary distinctiveness for trees in a \code{consInt} object
}
\examples{
example(make.consInt)
temp <- eDisCalc(combo.quercus)
plot(all.ed.equalSplits ~ sub.ed.equalSplits, temp$intersect)
}
