% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.consInt.R
\name{make.consInt}
\alias{make.consInt}
\title{Make a \code{consInt} object}
\usage{
make.consInt(phy, dat, matchCol = NA, imposeTidy = TRUE, ...)
}
\arguments{
\item{phy}{a \code{phylo} object}

\item{dat}{a \code{data.frame} object}

\item{matchCol}{the column in \code{dat} to match to the tree; if not provided,
\code{make.consInt} will use row names}

\item{imposeTidy}{\code{boolean} -- forces trees to match data using tidyName}

\item{...}{extra parameters to pass to \code{tidyName}}
}
\description{
\code{make.consInt} welds together a \code{data.frame} with a \code{phylo} object to yield
an object that can be analyzed for most analyses. The builder keeps
the whole tree plus a tree subsetted to the table provided; and the whole
table plus a table subsetted to the rows in the tree
}
\examples{
### 1st block redundant with cleanPhylo example
data(malus_tr)
data(quercus_tr)
data(tilia_tr)
data(ulmus_tr)
malus2 <- cleanPhylo(malus, delim = '_|_')
quercus2 <- cleanPhylo(quercus)
tilia2 <- cleanPhylo(tilia)
ulmus2 <- ulmus
ulmus2$tip.label <- paste('Ulmus', ulmus2$tip.label)

### Not duplicated in cleanPhylo:
data(accessions.mor) # read MOR accessions data
data(desiderata.mor) # read MOR desiderata list
temp <- c(accessions.mor$Taxon, desiderata.mor$taxon_name) \%>\%
   unique \%>\% sort
dat.mor <- data.frame(inGarden = temp \%in\% accessions.mor$Taxon,
                      wanted = temp \%in\% desiderata.mor$taxon_name,
                      row.names = temp)
rm(temp)
combo.malus <- make.consInt(malus2, dat.mor)
combo.quercus <- make.consInt(quercus2, dat.mor)
# combo.tilia <- make.consInt(tilia2, dat.mor) ## duplicate somewhere
combo.ulmus <- make.consInt(ulmus2, dat.mor)

}
