\name{PDvED-package}
\alias{PDvED-package}
\alias{PDvED}
\docType{package}
\title{
\packageTitle{PDvED}
}
\description{
\packageDescription{PDvED}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PDvED}
\packageIndices{PDvED}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{PDvED}

Maintainer: \packageMaintainer{PDvED}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
