% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyName.R
\name{tidyName}
\alias{tidyName}
\title{Tidy up a name}
\usage{
tidyName(x, fixes = c("[_., -'\\"]"), makeLow = TRUE, grepFix = FALSE)
}
\arguments{
\item{x}{\code{character} object}

\item{fixes}{what to get rid of}

\item{makeLow}{\code{boolean} object; indicates whether to get rid of case}

\item{grepFix}{\code{boolean}; indicates whether \code{fixed = TRUE} in \code{gsub}}
}
\description{
Tidy up a name
}
