% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w.phy.R
\name{w.phy}
\alias{w.phy}
\title{Calculate phylogenetic distinctiveness}
\usage{
w.phy(phy, useSqrt = TRUE)
}
\arguments{
\item{phy}{\code{phylo} object}

\item{useSqrt}{\code{boolean} - return sqrt?}
}
\description{
Calculate phylogenetic distinctiveness
}
