% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanPhylo.R
\name{cleanPhylo}
\alias{cleanPhylo}
\title{Clean a phylogeny for single-tip-per-species analyses}
\usage{
cleanPhylo(
  tr,
  delim = "|",
  tipElem = 1,
  uniquesOnly = TRUE,
  strFixed = TRUE,
  gsubFromTo = c("_", " "),
  gsubFixed = TRUE
)
}
\arguments{
\item{tr}{A \code{phylo} object}

\item{delim}{Character delimiting species names from other info}

\item{tipElem}{Integer indicating which element of the tip name contains the species}

\item{uniquesOnly}{Boolean: retain only unique tip labels? Generally this is what you want}

\item{strFixed}{Boolean: use \code{fixed = TRUE} in \code{strsplit}?}

\item{gsubFromTo}{Character, \code{length = 2}; or \code{NA}: if character, use with \code{gsub} to clean tip labels}

\item{gsubFixed}{Boolean: use \code{fixed = TRUE} in \code{gsub}?}
}
\description{
Clean a phylogeny for single-tip-per-species analyses
}
\examples{
data(malus.tr)
data(quercus.tr)
data(tilia.tr)
data(ulmus.tr)
malus2 <- cleanPhylo(malus.tr, delim = '_|_')
quercus2 <- cleanPhylo(quercus.tr)
tilia2 <- cleanPhylo(tilia.tr)
ulmus2 <- ulmus.tr
ulmus2$tip.label <- paste('Ulmus', ulmus2$tip.label)
}
